+++ packages/orq-rc/src/orq_ai_sdk/models/updatedatasetop.py
@@ -121,5 +121,5 @@ class UpdateDatasetResponseBody(BaseModel):
     created: Optional[datetime] = None
     r"""The date and time the resource was created"""
 
-    updated: Optional[datetime] = parse_datetime("2025-11-28T05:14:49.673Z")
+    updated: Optional[datetime] = parse_datetime("2025-11-28T12:16:10.264Z")
     r"""The date and time the resource was last updated"""