+++ packages/orq-rc/src/orq_ai_sdk/models/updatetoolop.py
@@ -188,7 +188,7 @@ class RequestBodyTools(BaseModel):
 
     schema_: Annotated[UpdateToolRequestBodyToolsSchema, pydantic.Field(alias="schema")]
 
-    id: Optional[str] = "01KB4E5QHHJSM840BNHY5R6NAH"
+    id: Optional[str] = "01KB569710CJKTR3Z877523MVS"
 
     description: Optional[str] = None
 
@@ -911,7 +911,7 @@ class UpdateToolResponseBodyCodeExecutionTool(BaseModel):
     code_tool: UpdateToolResponseBodyCodeTool
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "tool_01KB4E5QH5XT2H7B84MFHMQ01Z"
+        "tool_01KB56970S4NH4CV6MFA2SE06B"
     )
 
     display_name: Optional[str] = None
@@ -985,7 +985,7 @@ class UpdateToolResponseBodyTools(BaseModel):
         UpdateToolResponseBodyToolsSchema, pydantic.Field(alias="schema")
     ]
 
-    id: Optional[str] = "01KB4E5QH47K8ZDQZATDR4K6V1"
+    id: Optional[str] = "01KB56970RNCZ9THQHGR04TJMQ"
 
     description: Optional[str] = None
 
@@ -1084,7 +1084,7 @@ class UpdateToolResponseBodyMCPTool(BaseModel):
     mcp: UpdateToolResponseBodyMcp
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "tool_01KB4E5QH20F56RFG9ZR2EY70Y"
+        "tool_01KB56970PP6M8EMXFXAVXR3TX"
     )
 
     display_name: Optional[str] = None
@@ -1296,7 +1296,7 @@ class UpdateToolResponseBodyHTTPTool(BaseModel):
     http: UpdateToolResponseBodyHTTP
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "tool_01KB4E5QGZEHW5PCRNMW4EB0SD"
+        "tool_01KB56970KA308F69AS6GTNYXW"
     )
 
     display_name: Optional[str] = None
@@ -1450,7 +1450,7 @@ class UpdateToolResponseBodyJSONSchemaTool(BaseModel):
     json_schema: UpdateToolResponseBodyJSONSchema
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "tool_01KB4E5QGXQHHMYSMXNY1JGQRC"
+        "tool_01KB56970HG8H41TCSMTEETE4Z"
     )
 
     display_name: Optional[str] = None
@@ -1608,7 +1608,7 @@ class UpdateToolResponseBodyFunctionTool(BaseModel):
     function: UpdateToolResponseBodyFunction
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "tool_01KB4E5QGT3HW4QMW77ECBDGCN"
+        "tool_01KB56970GM48Q3QFP72AJVN9J"
     )
 
     display_name: Optional[str] = None